feat: GCC-009 Update Mutation  - Create update mutation  - Organize files, imports and exports
