+++ schema.js
@@ -29,10 +29,15 @@ export const typeDefs = `#graphql
     type Mutation {
         addGame(game: AddGameInput): Game
         deleteGame(id: ID!): [Game]
+        updateGame(id: ID!, edits: EditGameInput!): Game
     }
     input AddGameInput {
         title: String!,
         platform: [String!]!
+    }
+    input EditGameInput {
+        title: String,
+        platform: [String!]
     }
 `;
 // int, float, string, boolean, ID
